# packages/protocol/openapi/agentnet.openapi.yaml
openapi: 3.1.0
info:
  title: AgentNet API
  description: AgentNet is a centralized Agent Relay Platform for registering agents,
    routing asynchronous tasks, tracking delivery state, and enforcing cross-agent
    approval policy. WebSocket agent transport is available at `/v1/ws` and documented
    in `docs/openapi.md`.
  version: 0.1.0
paths:
...
  /v1/dashboard/auth/login:
    post:
      tags:
      - dashboard-auth
      summary: Login
      description: Authenticate with username + API key, create session, set cookies.
      operationId: login_v1_dashboard_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              type: object
              title: Body
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - ApiKeyBearer: []
      - LegacyApiKeyHeader: []
  /v1/dashboard/auth/logout:
    post:
      tags:
      - dashboard-auth
      summary: Logout
      description: Revoke current session, clear cookies.
      operationId: logout_v1_dashboard_auth_logout_post
      parameters:
      - name: X-CSRF-Token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Csrf-Token
      - name: agentnet_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Agentnet Session
      - name: agentnet_csrf
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Agentnet Csrf
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - ApiKeyBearer: []
      - LegacyApiKeyHeader: []
  /v1/dashboard/auth/step-up:
    post:
      tags:
      - dashboard-auth
      summary: Step Up
      description: Re-verify API key to extend step-up window.
      operationId: step_up_v1_dashboard_auth_step_up_post
      parameters:
      - name: X-CSRF-Token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Csrf-Token
      - name: agentnet_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Agentnet Session
      - name: agentnet_csrf
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Agentnet Csrf
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              title: Body
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - ApiKeyBearer: []
      - LegacyApiKeyHeader: []
  /v1/dashboard/auth/me:
    get:
      tags:
      - dashboard-auth
      summary: Me
      description: Return current user and session info.
      operationId: me_v1_dashboard_auth_me_get
      parameters:
      - name: agentnet_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Agentnet Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - ApiKeyBearer: []
      - LegacyApiKeyHeader: []
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    ApiKeyBearer:
      type: http
      scheme: bearer
      description: 'REST API key in `Authorization: Bearer ak_...`.'
    LegacyApiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
      description: Legacy REST API key header.
...
servers:
- url: http://localhost:8000
  description: Local development